add LandingPage.js
